# src/main/java/com/eugeniaArias/spring_security_jwt/docs/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/api/v1","description":"Generated server url"}],"paths":{"/users/updatePassword/{id}":{"put":{"tags":["user-controller"],"operationId":"updatePassword","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/users/createWithRoles":{"post":{"tags":["user-controller"],"operationId":"createUserWithRoles","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserNotPasswordResponseDto"}}}}}}},"/roles/createWithPermissions":{"post":{"tags":["role-controller"],"operationId":"createRoleWithPermissions","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleWithPermissionsRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponseDto"}}}}}}},"/roles/create/{roleName}":{"post":{"tags":["role-controller"],"operationId":"createRole","parameters":[{"name":"roleName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponseDto"}}}}}}},"/permissions/createBasicPermissions":{"post":{"tags":["permission-controller"],"operationId":"createBasicPermissions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/permissions/create/{permissionName}":{"post":{"tags":["permission-controller"],"operationId":"create","parameters":[{"name":"permissionName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PermissionResponseDto"}}}}}}},"/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserNotPasswordResponseDto"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthLoginResponseDto"}}}}}}},"/users/update/{id}":{"patch":{"tags":["user-controller"],"operationId":"updateUSer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserNotPasswordResponseDto"}}}}}}},"/users/toggleUserStatus/{id}":{"patch":{"tags":["user-controller"],"operationId":"toggleUserStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/roles/removePermissions":{"patch":{"tags":["role-controller"],"operationId":"removePermission","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleWithPermissionsRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponseDto"}}}}}}},"/permissions/update/id/{id}/{newPermissionName}":{"patch":{"tags":["permission-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"newPermissionName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PermissionResponseDto"}}}}}}},"/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserNotPasswordResponseDto"}}}}}}}},"/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserNotPasswordResponseDto"}}}}}}},"/users/username/{username}":{"get":{"tags":["user-controller"],"operationId":"getByUsername","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserNotPasswordResponseDto"}}}}}}},"/users/active-count":{"get":{"tags":["user-controller"],"operationId":"countActiveUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/roles":{"get":{"tags":["role-controller"],"operationId":"getAllRoles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleResponseDto"}}}}}}}},"/roles/name/{roleName}":{"get":{"tags":["role-controller"],"operationId":"getByName","parameters":[{"name":"roleName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponseDto"}}}}}}},"/roles/id/{id}":{"get":{"tags":["role-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponseDto"}}}}}},"delete":{"tags":["role-controller"],"operationId":"deleteById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/roles/count":{"get":{"tags":["role-controller"],"operationId":"count","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/permissions":{"get":{"tags":["permission-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PermissionResponseDto"}}}}}}}},"/permissions/name/{permissionName}":{"get":{"tags":["permission-controller"],"operationId":"getByName_1","parameters":[{"name":"permissionName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PermissionResponseDto"}}}}}}},"/permissions/id/{id}":{"get":{"tags":["permission-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PermissionResponseDto"}}}}}},"delete":{"tags":["permission-controller"],"operationId":"deleteById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/users/delete/{id}":{"delete":{"tags":["user-controller"],"operationId":"deleteById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UpdatePasswordRequestDto":{"required":["newPassword","password"],"type":"object","properties":{"password":{"type":"string"},"newPassword":{"type":"string"}}},"CreateUserRequestDto":{"required":["password","rolesIds","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"rolesIds":{"uniqueItems":true,"type":"array","items":{"type":"integer","format":"int64"}}}},"UserNotPasswordResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"enabled":{"type":"boolean"},"accountNotExpired":{"type":"boolean"},"accountNotLocked":{"type":"boolean"},"credentialNotExpired":{"type":"boolean"},"roleNames":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"RoleWithPermissionsRequestDto":{"required":["permissionIds","roleName"],"type":"object","properties":{"roleName":{"type":"string"},"permissionIds":{"uniqueItems":true,"type":"array","items":{"type":"integer","format":"int64"}}}},"RoleResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roleName":{"type":"string"},"permissionsName":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"PermissionResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"permissionName":{"type":"string"}}},"AuthRegisterDto":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthLoginRequestDto":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthLoginResponseDto":{"type":"object","properties":{"username":{"type":"string"},"message":{"type":"string"},"token":{"type":"string"}}}}}}